+++ sdk/spring/azure-spring-cloud-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/resourcemanager/AzureServiceBusResourceManagerAutoConfiguration.java
@@ -5,8 +5,8 @@ package com.azure.spring.cloud.autoconfigure.resourcemanager;
 
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.cloud.autoconfigure.servicebus.AzureServiceBusProperties;
-import com.azure.spring.servicebus.provisioning.DefaultServiceBusQueueProvisioner;
-import com.azure.spring.servicebus.provisioning.DefaultServiceBusTopicProvisioner;
+import com.azure.spring.servicebus.provisioning.arm.DefaultServiceBusQueueProvisioner;
+import com.azure.spring.servicebus.provisioning.arm.DefaultServiceBusTopicProvisioner;
 import com.azure.spring.cloud.resourcemanager.connectionstring.ServiceBusArmConnectionStringProvider;
 import com.azure.spring.servicebus.provisioning.ServiceBusQueueProvisioner;
 import com.azure.spring.servicebus.provisioning.ServiceBusTopicProvisioner;
