+++ sdk/spring/azure-spring-eventhubs/src/main/java/com/azure/spring/eventhubs/provisioning/arm/DefaultEventHubProvisioner.java
@@ -1,13 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.eventhubs.provisioning;
+package com.azure.spring.eventhubs.provisioning.arm;
 
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.EventHubConsumerGroupCrud;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.EventHubCrud;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.EventHubNamespaceCrud;
 import com.azure.spring.core.properties.resource.AzureResourceMetadata;
+import com.azure.spring.eventhubs.provisioning.EventHubProvisioner;
 import reactor.util.function.Tuples;
 
 /**
