+++ sdk/spring/azure-spring-servicebus/src/main/java/com/azure/spring/servicebus/provisioning/arm/DefaultServiceBusQueueProvisioner.java
@@ -1,11 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.spring.servicebus.provisioning;
+package com.azure.spring.servicebus.provisioning.arm;
 
 import com.azure.resourcemanager.AzureResourceManager;
 import com.azure.spring.cloud.resourcemanager.implementation.crud.ServiceBusQueueCrud;
 import com.azure.spring.core.properties.resource.AzureResourceMetadata;
+import com.azure.spring.servicebus.provisioning.ServiceBusQueueProvisioner;
 import reactor.util.function.Tuples;
 
 /**
