+++ sdk/spring/azure-spring-servicebus/src/main/java/com/azure/spring/servicebus/provisioning/arm/package-info.java
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * Package com.azure.spring.servicebus.provisioning.arm
+ */
+package com.azure.spring.servicebus.provisioning.arm;
